+++ src/jadepool/jadepool.d.ts
@@ -32,6 +32,8 @@ declare type EnvObject = {
 	param?: string
 	/** 进程启动时指定的任务名 */
 	task?: string
+	/** 进程需要负责的任务内容，以,号分割 */
+	jobs?: string
 	/** 进程脚本文件名，用于启动新进程 */
 	script: string
 	/** 主机HOST */
@@ -46,8 +48,6 @@ declare type EnvObject = {
 	secret: string
 	/** 设置后worker将自动启动, 默认true */
 	autoStart: boolean
-	/** 设置是否支持多同链多Worker模式, 默认false */
-	multiWorkers: boolean
 }
 
 declare class JadepoolSingleton {
@@ -64,13 +64,16 @@ declare class JadepoolSingleton {
 	 * 服务发现
 	 */
 	public consulSrv: services.ConsulService
+	/**
+	 * 自动注册的服务
+	 */
+	public autoRegisterServices: string[]
   
 	/**
 	 * 注册服务
 	 * @param serviceClass 
 	 * @param opts 
 	 */
-	registerService(name: 'agenda', opts: services.AgendaOptions): Promise<services.AgendaService>;
 	registerService(name: 'express', opts: services.AppOptions): Promise<services.AppService>;
 	registerService(name: 'error.code', opts: services.ErrorCodeOptions): Promise<services.ErrorCodeService>;
 	registerService(name: 'script', opts: services.ScriptOptions): Promise<services.ScriptService>;
@@ -79,17 +82,18 @@ declare class JadepoolSingleton {
 	registerService(name: 'jsonrpc.internal', opts: services.InternalRPCOptions): Promise<services.InternalRpcService>;
 	registerService(name: 'socket.io', opts: services.SocketIOOptions): Promise<services.SocketIOService>;
 	registerService(name: 'socket.io.worker', opts: services.SocketIOWorkerOptions): Promise<services.SocketIOWorkerService>;
+	registerService(name: 'pm2.process', opts: any): Promise<services.Pm2Service>;
 	registerService(name: 'child.process', opts: any): Promise<services.ProcessService>;
 	registerService(name: 'async.plan', opts: services.AsyncPlanOptions): Promise<services.AsyncPlanService>;
 	registerService(name: 'config', opts: services.ConfigOptions): Promise<services.ConfigService>;
 	registerService(name: 'consul', opts: services.ConsulOptions): Promise<services.ConsulService>;
+	registerService(name: 'job.queue', opts: services.JobQueueOptions): Promise<services.JobQueueService>;
 
 	/**
 	 * 确保服务已注册
 	 * @param serviceClass 
 	 * @param opts 
 	 */
-	ensureService(name: 'agenda', opts: services.AgendaOptions): Promise<services.AgendaService>;
 	ensureService(name: 'express', opts: services.AppOptions): Promise<services.AppService>;
 	ensureService(name: 'error.code', opts: services.ErrorCodeOptions): Promise<services.ErrorCodeService>;
 	ensureService(name: 'script', opts: services.ScriptOptions): Promise<services.ScriptService>;
@@ -98,16 +102,17 @@ declare class JadepoolSingleton {
 	ensureService(name: 'jsonrpc.internal', opts: services.InternalRPCOptions): Promise<services.InternalRpcService>;
 	ensureService(name: 'socket.io', opts: services.SocketIOOptions): Promise<services.SocketIOService>;
 	ensureService(name: 'socket.io.worker', opts: services.SocketIOWorkerOptions): Promise<services.SocketIOWorkerService>;
+	ensureService(name: 'pm2.process', opts: any): Promise<services.Pm2Service>;
 	ensureService(name: 'child.process', opts: any): Promise<services.ProcessService>;
 	ensureService(name: 'async.plan', opts: services.AsyncPlanOptions): Promise<services.AsyncPlanService>;
 	ensureService(name: 'config', opts: services.ConfigOptions): Promise<services.ConfigService>;
 	ensureService(name: 'consul', opts: services.ConsulOptions): Promise<services.ConsulService>;
+	ensureService(name: 'job.queue', opts: services.JobQueueOptions): Promise<services.JobQueueService>;
 
 	/**
 	 * 获取服务
 	 * @param name
 	 */
-	getService(name: 'agenda'): services.AgendaService;
 	getService(name: 'express'): services.AppService;
 	getService(name: 'error.code'): services.ErrorCodeService;
 	getService(name: 'script'): services.ScriptService;
@@ -116,10 +121,12 @@ declare class JadepoolSingleton {
 	getService(name: 'jsonrpc.internal'): services.InternalRpcService;
 	getService(name: 'socket.io'): services.SocketIOService;
 	getService(name: 'socket.io.worker'): services.SocketIOWorkerService;
+	getService(name: 'pm2.process'): services.Pm2Service;
 	getService(name: 'child.process'): services.ProcessService;
 	getService(name: 'async.plan'): services.AsyncPlanService;
 	getService(name: 'config'): services.ConfigService;
 	getService(name: 'consul'): services.ConsulService;
+	getService(name: 'job.queue'): services.JobQueueService;
 
 	/**
 	 * 获取模型