+++ src/services/asyncplan.service.js
@@ -1,10 +1,11 @@
 const _ = require('lodash')
 const BaseService = require('./core')
 const consts = require('../consts')
-const NBError = require('../NBError')
 const jadepool = require('../jadepool')
 
 const logger = require('@jadepool/logger').of('Service', 'Async Plan')
+
+const JOB_NAME = 'async-plan-service-tick'
 
 class Service extends BaseService {
   /**
@@ -12,11 +13,6 @@ class Service extends BaseService {
    */
   constructor (services) {
     super(consts.SERVICE_NAMES.ASYNC_PLAN, services)
-    // check required
-    const agendaSrv = jadepool.getService(consts.SERVICE_NAMES.AGENDA)
-    if (!agendaSrv) {
-      throw new NBError(10001, `missing agenda service`)
-    }
   }
 
   /**
@@ -25,21 +21,26 @@ class Service extends BaseService {
    * @param {number} [opts.processEvery=30] 检测间隔
    */
   async initialize (opts) {
-    const agendaSrv = jadepool.getService(consts.SERVICE_NAMES.AGENDA)
-    const agendaNative = agendaSrv._agenda
-    // 运行循环任务
-    const taskName = 'async-plan-service-tick'
-    agendaNative.define(taskName, { priority: 'high', concurrency: 1 }, async (job, done) => {
+    const jobSrv = await jadepool.ensureService(consts.SERVICE_NAMES.JOB_QUEUE)
+    const queue = await jobSrv.fetchQueue(JOB_NAME) // 默认队列
+    queue.process('every', async (job) => {
       try {
         await this._everyHandler()
       } catch (err) {
         logger.error(`unexpected`, err)
       }
-      // 完成任务
-      done()
     })
-    const sec = opts.processEvery || 30
-    await agendaNative.every(`${sec} seconds`, taskName)
+    logger.tag('JobQueue Registered').log(`name=${JOB_NAME}`)
+
+    const tickDelta = opts.processEvery || 30
+    // 运行循环任务
+    this._job = await queue.add('every', {}, {
+      priority: 1,
+      repeat: {
+        every: tickDelta * 1000
+      }
+    })
+    logger.tag('Initialized').log(`interval=${tickDelta},jobId=${this._job.id}`)
   }
 
   /**
@@ -47,7 +48,12 @@ class Service extends BaseService {
    * @param signal 退出信号
    */
   async onDestroy (signal) {
-    // NOTHING
+    if (this._job) {
+      this._job.remove()
+    }
+    const jobSrv = jadepool.getService(consts.SERVICE_NAMES.JOB_QUEUE)
+    const queue = await jobSrv.fetchQueue(JOB_NAME) // 默认队列
+    await queue.clean(0)
   }
 
   /**