+++ src/utils/index.js
@@ -1,6 +1,7 @@
 module.exports = {
   assert: require('./assert'),
   string: require('./string'),
+  waitForSeconds: require('./waitForSeconds'),
   config: require('./config/index'),
   configLoader: require('./config/loader'),
   configSetuper: require('./config/setuper'),