+++ src/utils/redis.d.ts
@@ -2,6 +2,6 @@ import redis from 'redis'
 
 export function initialize (): Promise<void>
 
-export function getOpts (): { url: string, [key: string]: string }
+export function getOpts (name?: string): { url: string, [key: string]: string }
 
 export function fetchClient (name: string): redis.RedisClient